requirements.txt
Cerberus==1.3.4
PyYAML==6.0.1
python-dateutil==2.8.2
semver==3.0.0.dev3
toml==0.10.2
typer==0.12.5
typing_extensions==4.8.0
xlsx2csv==0.8.3
